Refactor Tag management by restructuring resources
- Updated the namespace for TagResource to Firefly\FilamentBlog\Resources\Tags.
- Removed the TagResource class and its associated pages (CreateTag, EditTag, ListTags) to streamline the codebase.
- Eliminated the deprecated getForm method from the Tag model to enhance clarity and maintainability.
- Adjusted Blog.php to reflect the new TagResource path.
