+++ src/Models/Tag.php
@@ -2,13 +2,10 @@
 
 namespace Firefly\FilamentBlog\Models;
 
-use Filament\Schemas\Components\Utilities\Set;
-use Filament\Forms\Components\TextInput;
 use Firefly\FilamentBlog\Database\Factories\TagFactory;
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
-use Illuminate\Support\Str;
 
 class Tag extends Model
 {
@@ -27,25 +24,6 @@ class Tag extends Model
     {
 
         return $this->belongsToMany(Post::class, config('filamentblog.tables.prefix').'post_'.config('filamentblog.tables.prefix').'tag');
-    }
-
-    public static function getForm(): array
-    {
-        return [
-            TextInput::make('name')
-                ->live(true)->afterStateUpdated(fn(Set $set, ?string $state) => $set(
-                    'slug',
-                    Str::slug($state)
-                ))
-                ->unique(config('filamentblog.tables.prefix').'tags', 'name', null, 'id')
-                ->required()
-                ->maxLength(50),
-
-            TextInput::make('slug')
-                ->unique(config('filamentblog.tables.prefix').'tags', 'slug', null, 'id')
-                ->readOnly()
-                ->maxLength(155),
-        ];
     }
 
     protected static function newFactory()
