--- src/Resources/TagResource.php
@@ -1,82 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources;
-
-use Filament\Schemas\Schema;
-use Filament\Tables\Columns\TextColumn;
-use Filament\Actions\EditAction;
-use Filament\Actions\DeleteAction;
-use Filament\Actions\ViewAction;
-use Filament\Actions\BulkActionGroup;
-use Filament\Actions\DeleteBulkAction;
-use Firefly\FilamentBlog\Resources\TagResource\Pages\ListTags;
-use Firefly\FilamentBlog\Resources\TagResource\Pages\EditTag;
-use Filament\Resources\Resource;
-use Filament\Tables;
-use Filament\Tables\Table;
-use Firefly\FilamentBlog\Models\Tag;
-
-class TagResource extends Resource
-{
-    protected static ?string $model = Tag::class;
-
-    protected static string | \BackedEnum | null $navigationIcon = 'heroicon-o-tag';
-
-    protected static string | \UnitEnum | null $navigationGroup = 'Blog';
-
-    protected static ?int $navigationSort = 2;
-
-    public static function form(Schema $schema): Schema
-    {
-        return $schema
-            ->components(Tag::getForm());
-    }
-
-    public static function table(Table $table): Table
-    {
-        return $table
-            ->columns([
-                TextColumn::make('name')
-                    ->searchable(),
-                TextColumn::make('slug'),
-
-                TextColumn::make('created_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-
-                TextColumn::make('updated_at')
-                    ->dateTime()
-                    ->sortable()
-                    ->toggleable(isToggledHiddenByDefault: true),
-            ])
-            ->filters([
-                //
-            ])
-            ->recordActions([
-                EditAction::make(),
-                DeleteAction::make(),
-                ViewAction::make(),
-            ])
-            ->toolbarActions([
-                BulkActionGroup::make([
-                    DeleteBulkAction::make(),
-                ]),
-            ]);
-    }
-
-    public static function getRelations(): array
-    {
-        return [
-            //
-        ];
-    }
-
-    public static function getPages(): array
-    {
-        return [
-            'index' => ListTags::route('/'),
-            'edit' => EditTag::route('/{record}/edit'),
-        ];
-    }
-}
