--- src/Resources/TagResource/Pages/CreateTag.php
@@ -1,11 +1,0 @@
-<?php
-
-namespace Firefly\FilamentBlog\Resources\TagResource\Pages;
-
-use Filament\Resources\Pages\CreateRecord;
-use Firefly\FilamentBlog\Resources\TagResource;
-
-class CreateTag extends CreateRecord
-{
-    protected static string $resource = TagResource::class;
-}
